requirements.txt
asyncio
logging
websockets
jsonrpcclient
openai
python-dotenv
beautifulsoup4
requests
